Tercer testing PASSED

+++ app/calculator.py
@@ -9,7 +9,7 @@ class Calculator:
         'division': '/'
     }
 
-    def get_expression(self, num1: int, num2: int, operation: str) -> str:
+    def get_expression(self, num1: float, num2: float, operation: str) -> str:
         try:
             opr: str = self._OPERATIONS[operation]
         except KeyError:
@@ -18,11 +18,11 @@ class Calculator:
         expression = f'{num1}{opr}{num2}'
         return expression
 
-    def calculate(self, num1: int, num2: int, operation: str) -> int:
+    def calculate(self, num1: float, num2: float, operation: str) -> int:
         expression = self.get_expression(num1, num2, operation)
 
         try:
             result: float = eval(expression)
-            return int(result)
+            return float(result)
         except Exception as e:
             raise MathematicalException(f'The mathematical operation is not correct: {e}')
